pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitcompass"
version = "0.1.0"
description = "A powerful Python-based tool for managing GitHub projects, issues, sub-issues, and roadmaps"
authors = [{name = "PimpMyNines", email = "team@pimpmy9s.com"}]
requires-python = ">=3.8"
dependencies = [
    "pygithub>=1.58.0",
    "click>=8.1.3",
    "pyyaml>=6.0",
    "requests>=2.28.1",
    "python-dotenv>=0.21.0",
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Version Control :: Git",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "flake8>=6.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "build>=0.10.0",
]

[project.scripts]
gitcompass = "octomaster.cli:main"

[project.urls]
Homepage = "https://github.com/PimpMyNines/github-project-management"
Repository = "https://github.com/PimpMyNines/github-project-management.git"
Issues = "https://github.com/PimpMyNines/github-project-management/issues"
Documentation = "https://github.com/PimpMyNines/github-project-management/tree/main/docs"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
